refactor: move style helpers to @lwc/shared
