+++ packages/@lwc/engine-core/src/framework/utils.ts
@@ -67,28 +67,6 @@ export function guid(): string {
     return s4() + s4() + '-' + s4() + '-' + s4() + '-' + s4() + '-' + s4() + s4() + s4();
 }
 
-// Borrowed from Vue template compiler.
-// https://github.com/vuejs/vue/blob/531371b818b0e31a989a06df43789728f23dc4e8/src/platforms/web/util/style.js#L5-L16
-const DECLARATION_DELIMITER = /;(?![^(]*\))/g;
-const PROPERTY_DELIMITER = /:(.+)/;
-
-export function parseStyleText(cssText: string): { [name: string]: string } {
-    const styleMap: { [name: string]: string } = {};
-
-    const declarations = cssText.split(DECLARATION_DELIMITER);
-    for (const declaration of declarations) {
-        if (declaration) {
-            const [prop, value] = declaration.split(PROPERTY_DELIMITER);
-
-            if (prop !== undefined && value !== undefined) {
-                styleMap[prop.trim()] = value.trim();
-            }
-        }
-    }
-
-    return styleMap;
-}
-
 // Make a shallow copy of an object but omit the given key
 export function cloneAndOmitKey(object: { [key: string]: any }, keyToOmit: string) {
     const result: { [key: string]: any } = {};
