+++ packages/@lwc/template-compiler/src/index.ts
@@ -25,7 +25,6 @@ export { Config } from './config';
 export { toPropertyName } from './shared/utils';
 export { kebabcaseToCamelcase } from './shared/naming';
 export { generateScopeTokens } from './scopeTokens';
-export { normalizeStyleAttribute } from './codegen/helpers';
 
 /**
  * Parses HTML markup into an AST
